Add files via upload
Create Table statements for MySQL Workbench.
I made Table Statements for User, Page, System, Admin, Developer, Game, and Library. I was unsure on what the Like and Comment tables should be so I did not create them. Feel free to make changes to the statements if unclear or broken for what we need for the Back End.